Added static pod generation functionality
A new function, generateStaticPod, has been added to the remote.go file. This function creates a static pod manifest on a remote system. It takes in parameters such as connection details, filename and pod manifest and returns outputs like the remote manifest file and its path. The changes also include adding 'StaticPod' to the list of commands that can be generated remotely.