Added connection function in remote utilities
A new function named 'connection' has been added to the remote utilities. This function returns a PropertySpec object that describes the parameters needed to connect to a remote host. The returned object includes a description and a TypeSpec, which is an extended type from the command specification with "Connection" as its name and "remote" as its context.