Add etcd service (#112)
* Update go sdk deps

* Update kubernetes dep after renovate

* Refactor schemagen into smaller modules

* Fix whatever the fuck happened here

* Actually use the gen function

* Added connection function in remote utilities

A new function named 'connection' has been added to the remote utilities. This function returns a PropertySpec object that describes the parameters needed to connect to a remote host. The returned object includes a description and a TypeSpec, which is an extended type from the command specification with "Connection" as its name and "remote" as its context.

* Added etcdService to remote package

A new function, generateEtcdService, has been added to the remote package. This function generates a schema for an Etcd service with specific input and output properties. The generated schema is then added to the list of resources in the Generate function of the same package.

* WIP

* Added Integer function to property spec

A new function, Integer, has been added to the property specifications. This function takes a description as an argument and returns a schema.PropertySpec with the provided description and type set as integer.

* Updated etcdService.go with map functionality

The etcdService.go file has been updated to include the "maps" package. The 'restartSec' property type has been changed from string to integer. The 'ip' output property has been removed and a copy function from the maps package is used to copy inputs into outputs. The required inputs have also been updated, removing 'ip'.

* Added new EtcdService component

A new component, EtcdService, has been added to the schema. This includes properties for connection parameters, systemd unit description and documentation, restart behavior and timing, remote systemd service details, and wanted-by override. The connection property is required for this component.

* Added EtcdService to Kubernetes provider

A new component, EtcdService, has been added to the Kubernetes provider. This includes the addition of a new ResourceConstructor and an abstract class with its associated arguments interface. The EtcdService class extends pulumi.ComponentResource and contains several properties including connection, description, documentation, restart, restartSec, service and wantedBy.

* Added EtcdService class for remote host connection

This update introduces the EtcdService class which is responsible for managing connections to a remote host. The class includes options to specify systemd unit description, documentation, restart behaviour and more. This addition will enhance the flexibility and control over remote host connections in our application.

* Fixing things

* Regen schema

* Regen schema types

* Regen SDKs

* Implementing

* Add test resource

* Update schema

* Update schema types

* WIP trying to get it working

* Regen SDKs

* Update test

* Schema make target

* Etcd props type

* Regen schema

* Regen schema types

* Regen SDKs

* Attempt to implement

* Accept props as input

* Regen schema

* Regen schema types

* Regen SDK

* Implement

* Peer configuration

* Regen schema

* Regen code

* Implementing

* Fix the test