Add the etcd cluster resource (#116)
* Add etcd cluster

* Regen code

* Refactor etcdCluster function and update node properties

The etcdCluster generation function has been refactored to no longer require a commandSpec parameter. The 'nodes' property within the same function has also been updated to include a reference to the etcdNode type instead of directly using it. Additionally, an 'architecture' property has been added to the Etcd node description in types.go, which specifies the CPU architecture of the node.

* Regen code

* Implementing