Add more tools (#113)
* Refactor schemagen tools

* Sed tool

* Hostnamectl tool

* Start on curl

* Regen code

* WIP implementing sed

* Set command output

* Fix some schema stuff

* Regen code

* Implementing

* Add to resources

* Fix the thing

* Regen code

* Updated type of stripComponents

The type of the 'stripComponents' variable has been updated to allow for undefined values. This change provides more flexibility in handling cases where 'stripComponents' might not be defined.

* Updated type definitions and handling in commandBuilder and sed

Expanded the OptValue type definition in commandBuilder to include 'number' along with 'boolean' and 'string'. In sed, removed default value for lineLength argument and updated its type to allow undefined. This change provides more flexibility in handling different types of option values.