Etcd tests (#151)
* Remote etcd cluster test

* Fix some schema stuff

* Fixing bits and bobs

* Add new etcd install test

* Update etcdctl required inputs

* Add version command

* Fix test bug

* Update test

* Etcdinstall assertions

* Refactor etcd configuration and service

Removed the function formatSystemdFile from etcdConfiguration.ts. This function was responsible for formatting systemd files, but it's no longer needed. Also, in etcdService.ts, updated the --listen-client-urls option to only use clientUrl instead of both clientUrl and localhostUrl. The change simplifies the code and makes it more efficient by reducing unnecessary operations.

* Added Exec method to SshServer

The new Exec method allows execution of commands on the SSH server. It takes a context and command as parameters, executes the command, and returns the result as a string. Error handling has been implemented for both command execution and reading the result.

* Updated internal IP in etcd-cluster

The internal IP address for the 'node0' in the etcd-cluster has been updated from '10.69.0.2' to '0.0.0.0'. This change is part of an effort to generalize the configuration settings for broader use cases.

* Working test?!?

* More assertions

* Simplified test configuration

Removed unnecessary parameters from the test configuration in two functions. The 'Quick', 'SkipRefresh', and 'RunUpdateTest' options were removed, leaving only the 'Dir' option. This simplifies the setup and makes it easier to understand.

* Disable etcd install update test

* Taking the lazy approach