Updated test paths and added new plugin configuration
The commit includes changes to the test paths in the remote testing file. The path for bridge and loopback configurations have been updated. A new function, validatePlugins, has been added to check the outputs of a resource.

In addition, a new CniPluginConfiguration instance has been created in the index.ts file. This instance is dependent on 'mkdir' and includes connection details as well as subnet and directory information.