Fix plugin paths and add grouping resource (#243)
* GAH regen code

* Added CNI plugin configuration generation

A new function for generating CNI plugin configurations has been added. This includes the creation of input and output properties, as well as required inputs and outputs. The function is now also included in the package's index file.

* Regen code

* Refactor CNI plugin configuration

Removed the 'file' resource from the CNI plugin configuration and its reference in required outputs. This simplifies the code by eliminating unnecessary elements, enhancing readability and maintainability.

* Updated CNI plugin configuration

The CNI plugin configuration has been updated to include a new 'directory' property. This replaces the previous '/etc/cni/net.d' mkdir operation with a more generic 'mkdir' operation that works on the specified directory. The required inputs and outputs have also been adjusted accordingly.

* Regen code

* Add the new resource and fix paths

* Updated test paths and added new plugin configuration

The commit includes changes to the test paths in the remote testing file. The path for bridge and loopback configurations have been updated. A new function, validatePlugins, has been added to check the outputs of a resource.

In addition, a new CniPluginConfiguration instance has been created in the index.ts file. This instance is dependent on 'mkdir' and includes connection details as well as subnet and directory information.