Onboard the remote tests to the new helper function (#251)
* Onboard the etcd install test

* Refactor resource test name generation

The format of the test names in the resource tests has been updated. Previously, it was simply "Testing resource [resource name]". Now, it includes both the resource type and its name for better clarity and specificity during testing.

* Updated resource test name formatting

The format of the test names in the resource tests has been updated. Previously, the type and name were concatenated directly. Now, they are separated by "::" for better readability and organization.

* Refactored test for remote etcd cluster

The test for the remote etcd cluster has been significantly refactored. The previous implementation was replaced with a more streamlined version using ResourceTest and Validate functions. This change simplifies the code, making it easier to read and maintain.

* Refactored test functions in remote test file

Significant changes were made to the testing functions in the remote test file. The validation of bridge, loopback, and plugins was simplified by using a new function 'expectOutput'. This change reduces redundancy and improves readability. Also, the structure of tests has been streamlined for better organization.