Implement containerd configuration (#254)
* Way too many types

* Regen code

* Refactored type names in schema generation

The type names used in the schema generation code have been refactored for clarity and consistency. This includes changes to the naming of containerd configuration types, cni configuration types, and runc plugin configuration types. The changes ensure that the names accurately reflect their respective functionalities.

And also it was just a typo - Erik

* Refactored local resource and type functions

The local resource and type functions have been refactored for better code readability. The previously private functions 'localResource' and 'localType' are now public, renamed to 'LocalResourceRef' and 'LocalTypeRef'. These changes improve the modularity of the code by allowing external access to these functionalities.

* Updated type specifications in configuration

The type specifications for 'cri', 'containerd', 'cni', and 'options' in the configuration have been updated. The changes involve replacing the previous LocalType method with a TypeSpec structure that includes a Plain attribute set to true, and a Ref attribute using the LocalTypeRef method. This modification enhances code readability and maintainability.

* Regen code

* WIP on implementation

* Add new resource to resource map

* Add path input

* Regen code

* Include runtimes property

* Regen code

* Refactor defaults and format as toml

* Fix path + chores

* Added remote worker test

A new test function, TestRemoteWorkerTs, has been added to the examples_remote_test.go file. This function tests the 'remote/worker-ts' resource and validates its outputs. It checks for non-empty outputs and verifies that they contain "connection", "cri", and "file". The expected output path is also asserted.

* Add a worker ts program