Add controlplane node resource (#278)
* Bootstrappin

* Use the props thing

* WIP on the controlplane node

* More props

* Regen code

* More stuff for the resource

* Regen code

* Implementing the thing

* Include in resources map

* Regen code

* Missing dep

* Refactor worker node imports

Removed unnecessary import from the worker node file. This simplifies the code and reduces dependencies, making it easier to maintain in the future.

* Add new controlplane test

* Refactor control plane node class

Updated the ControlPlaneNode class to use instance variables for install directories. This change improves code readability and maintainability by ensuring that the correct directory paths are always used.

* Add assertions