Updated default containerd version and binaries list (#344)
* Updated default containerd version and binaries list

The default version of containerd has been updated from '1.4.4' to '1.7.19'. Additionally, the list of binaries for installation has been expanded to include 'containerd-shim', 'containerd-shim-runc-v1', 'containerd-shim-runc-v2', 'containerd-stress', and 'ctr'.

* Added triggers to archive and binary install functions

In the `archiveInstall` and `binaryInstall` functions, added triggers for URL and mktemp respectively. This ensures that these operations are triggered when the specified conditions are met, improving the efficiency of our code execution.