Updated systemctl tool implementation
The changes in the code include an update to the import statement, removing 'output' from '@pulumi/pulumi'. The builder object has been modified to handle 'opts.unit' and a new option '--now' has been added. A new property 'now' is also mapped in the return object of the factory function.