Added 'now' option to systemctl tool (#345)
* Added 'now' option to systemctl tool

A new property, 'now', has been added to the systemctl tool. This corresponds to the '--now' option in the command line interface.

* Regen code

* Regen schema types

* Updated systemctl tool implementation

The changes in the code include an update to the import statement, removing 'output' from '@pulumi/pulumi'. The builder object has been modified to handle 'opts.unit' and a new option '--now' has been added. A new property 'now' is also mapped in the return object of the factory function.