Update the default directory for installs (#348)
* Updated default directory in archiveInstall

The default installation directory has been updated based on the type of install. If the type is "CniPluginsInstall", it defaults to "/opt/cni/bin". Otherwise, it defaults to "/usr/local/bin". This change improves flexibility for different types of installations.

* Regen code

* Updated package.json files

The package.json files in both the 'pulumi-resource-kubernetes-the-hard-way' and 'provider-scripts' directories have been updated. The license field has been repositioned, and a new field for specifying the package manager along with its version has been added. This change ensures consistent use of the same package manager across different parts of the project.